+++ plugins/FTPFileYM/curl/lib/telnet.c
@@ -972,7 +972,7 @@ static void suboption(struct connectdata *conn)
       for(v = tn->telnet_vars;v;v = v->next) {
         tmplen = (strlen(v->data) + 1);
         /* Add the variable only if it fits */
-        if(len + tmplen < (int)sizeof(temp)-6) {
+        if(len + tmplen < sizeof(temp)-6) {
           sscanf(v->data, "%127[^,],%127s", varname, varval);
           snprintf((char *)&temp[len], sizeof(temp) - len,
                    "%c%s%c%s", CURL_NEW_ENV_VAR, varname,
